refactor: Migrate to ESM. Move controllers to submodules.
