--- index.js
@@ -1,2 +1,0 @@
-const AccessControllers = require('./src/access-controllers')
-module.exports = AccessControllers
