+++ package.json
@@ -6,9 +6,21 @@
     "orbit",
     "orbitdb"
   ],
-  "main": "index.js",
+  "type": "module",
+  "main": "./src/index.js",
+  "exports": {
+    ".": "./src/index.js",
+    "./contract": "./src/access-controllers/contract.js",
+    "./deposit-contract": "./src/access-controllers/deposit-contract.js",
+    "./interface": "./src/access-controllers/interface.js",
+    "./ipfs": "./src/access-controllers/ipfs.js",
+    "./legacy-ipfs": "./src/access-controllers/legacy-ipfs.js",
+    "./manifest": "./src/access-controllers/manifest.js",
+    "./orbitdb": "./src/access-controllers/orbitdb.js"
+  },
   "scripts": {
-    "lint": "standard"
+    "lint": "standard",
+    "lint:fix": "standard --fix"
   },
   "repository": {
     "type": "git",
@@ -27,31 +39,15 @@
   },
   "homepage": "https://github.com/orbitdb/orbit-db-access-controllers#readme",
   "dependencies": {
-    "p-map-series": "^1.0.0"
-  },
-  "devDependencies": {
-    "@colony/purser-software": "^1.1.2",
-    "eslint": "^7.32.0",
-    "ganache-cli": "^6.12.2",
-    "go-ipfs": "^0.10.0-rc1",
-    "ipfs": "^0.58.3",
-    "ipfsd-ctl": "^10.0.3",
-    "mocha": "^5.2.0",
-    "node-pre-gyp": "^0.13.0",
-    "orbit-db-identity-provider": "^0.4.0",
-    "orbit-db-io": "^1.0.1",
-    "orbit-db-keystore": "^1.0.0",
-    "orbit-db-test-utils": "^1.3.0",
-    "standard": "^14.1.0",
-    "web3": "^1.5.2"
-  },
-  "standard": {
-    "env": [
-      "mocha"
-    ]
+    "orbit-db-io": "github:haydenyoung/orbit-db-io#esm",
+    "p-map-series": "^1.0.0",
+    "safe-buffer": "^5.2.1"
   },
   "localMaintainers": [
     "shamb0t <shams@haja.io>",
     "hajamark <mark@haja.io>"
-  ]
+  ],
+  "devDependencies": {
+    "standard": "^17.0.0"
+  }
 }
