+++ src/utils/ensure-ac-address.js
@@ -1,10 +1,9 @@
-'use strict'
-const path = require('path')
+import path from 'path'
+
 // Make sure the given address has '/_access' as the last part
-const ensureAddress = address => {
+export default address => {
   const suffix = address.toString().split('/').pop()
   return suffix === '_access'
     ? address
     : path.join(address, '/_access')
 }
-module.exports = ensureAddress
