+++ src/utils/get-entry-key.js
@@ -1,5 +1,1 @@
-'use strict'
-
-const getEntryKey = (e) => e.v === 0 ? e.key : e.identity.publicKey
-
-module.exports = getEntryKey
+export default (e) => e.v === 0 ? e.key : e.identity.publicKey
