+++ src/utils/is-valid-eth-address.js
@@ -1,6 +1,3 @@
-'use strict'
-
-const isValidEthAddress = (web3, address) => {
+export default (web3, address) => {
   return web3.utils.isAddress(address)
 }
-module.exports = isValidEthAddress
